in int

+++ zakozhurnikova.kristina/S5/getSum.cpp
@@ -2,8 +2,8 @@
 
 void zakozhurnikova::KeySum::operator()(int data)
 {
-  long long maxNum = std::numeric_limits< long long >::max();
-  long long minNum = std::numeric_limits< long long >::min();
+  int maxNum = std::numeric_limits< int >::max();
+  int minNum = std::numeric_limits< int >::min();
   if ((result_ > 0 && maxNum - result_ < data) || (result_ < 0 && data < 0 && minNum - result_ > data))
   {
     throw std::overflow_error("Error: Addition overflow");
